Update prediction form to use integer (int) fields — input handling and Pydantic schema are now consistent.

+++ Server/app/schemas/prediction.py
@@ -2,12 +2,12 @@ from datetime import datetime
 from pydantic import BaseModel
 
 class InputRow(BaseModel):
-    Confirmed: float
-    Deaths: float
-    Recovered: float
-    Active: float
-    New_cases: float
-    New_recovered: float
+    Confirmed: int
+    Deaths: int
+    Recovered: int
+    Active: int
+    New_cases: int
+    New_recovered: int
     date: datetime
     Country: str
     WHO_Region: str
